Refactor MSSQL registry handlers into domain helpers
