+++ tests/test_users_session_upsert.py
@@ -1,88 +1,39 @@
-import asyncio, importlib.util, pathlib, sys, types
+import asyncio
 from contextlib import asynccontextmanager
 from datetime import datetime, timezone
-import server.modules.providers.database.mssql_provider  # ensure provider module loaded
 
-root_path = pathlib.Path(__file__).resolve().parent.parent
-
-# Stub package structure
-server_pkg = types.ModuleType("server")
-server_pkg.__path__ = [str(root_path / "server")]
-sys.modules.setdefault("server", server_pkg)
-modules_pkg = types.ModuleType("server.modules")
-modules_pkg.__path__ = [str(root_path / "server/modules")]
-sys.modules.setdefault("server.modules", modules_pkg)
-providers_pkg = types.ModuleType("server.modules.providers")
-providers_pkg.__path__ = [str(root_path / "server/modules/providers")]
-sys.modules.setdefault("server.modules.providers", providers_pkg)
-database_pkg = types.ModuleType("server.modules.providers.database")
-database_pkg.__path__ = [str(root_path / "server/modules/providers/database")]
-sys.modules.setdefault("server.modules.providers.database", database_pkg)
-mssql_pkg = types.ModuleType("server.modules.providers.database.mssql_provider")
-mssql_pkg.__path__ = [str(root_path / "server/modules/providers/database/mssql_provider")]
-sys.modules.setdefault("server.modules.providers.database.mssql_provider", mssql_pkg)
-
-# Stub dependencies
-a = types.ModuleType("server.modules.providers.database.mssql_provider.logic")
-sys.modules["server.modules.providers.database.mssql_provider.logic"] = a
-a.init_pool = lambda *args, **kwargs: None
-a.close_pool = lambda *args, **kwargs: None
-async def _dummy_tx():
-  yield
-
-a.transaction = lambda: _dummy_tx()
-
-b = types.ModuleType("server.modules.providers.database.mssql_provider.db_helpers")
-sys.modules["server.modules.providers.database.mssql_provider.db_helpers"] = b
-b.fetch_rows = lambda *args, **kwargs: None
-b.fetch_json = lambda *args, **kwargs: None
-b.exec_query = lambda *args, **kwargs: None
-b.Operation = type("Operation", (), {})
-b.exec_op = lambda *args, **kwargs: None
-b.json_many = lambda *args, **kwargs: None
-b.json_one = lambda *args, **kwargs: None
-b.row_many = lambda *args, **kwargs: None
-b.row_one = lambda *args, **kwargs: None
-
-spec = importlib.util.spec_from_file_location(
-  "server.modules.providers.database.mssql_provider.registry",
-  root_path / "server/modules/providers/database/mssql_provider/registry.py",
-)
-registry_mod = importlib.util.module_from_spec(spec)
-sys.modules["server.modules.providers.database.mssql_provider.registry"] = registry_mod
-spec.loader.exec_module(registry_mod)
+from server.registry.auth.session import mssql as auth_session
 
 
 def test_create_session_updates_existing(monkeypatch):
   executed: list[str] = []
-  fetch_calls: list[int] = []
   lookups: list[tuple[str, bool]] = []
 
   class DummyCur:
+    def __init__(self):
+      self._step = 0
     async def execute(self, sql, params):
       executed.append(sql.lower())
     async def fetchone(self):
-      if not fetch_calls:
-        fetch_calls.append(1)
-        return (1,)
-      if len(fetch_calls) == 1:
-        fetch_calls.append(1)
+      if self._step == 0:
+        self._step += 1
         return ("sess",)
-      return ("dev",)
+      if self._step == 1:
+        self._step += 1
+        return ("dev",)
+      return None
 
   @asynccontextmanager
   async def fake_tx():
     yield DummyCur()
 
-  monkeypatch.setattr(registry_mod, "transaction", fake_tx)
-  original_lookup = registry_mod.get_auth_provider_recid
-
-  async def fake_lookup(provider, cursor=None):
+  async def fake_lookup(provider, *, cursor=None):
     lookups.append((provider, cursor is not None))
-    return await original_lookup(provider, cursor=cursor)
+    return 1
 
-  monkeypatch.setattr(registry_mod, "get_auth_provider_recid", fake_lookup)
-  handler = registry_mod.get_handler("db:auth:session:create_session:1")
+  monkeypatch.setattr(auth_session, "transaction", fake_tx)
+  monkeypatch.setattr(auth_session, "get_auth_provider_recid", fake_lookup)
+
   args = {
     "access_token": "tok",
     "expires": datetime.now(timezone.utc),
@@ -92,7 +43,9 @@ def test_create_session_updates_existing(monkeypatch):
     "user_guid": "user",
     "provider": "microsoft",
   }
-  asyncio.run(handler(args))
+
+  asyncio.run(auth_session.create_session_v1(args))
+
   assert any("update users_sessions" in q for q in executed)
   assert not any("insert into users_sessions" in q for q in executed)
   assert any("update sessions_devices" in q for q in executed)
